- Solve issue #21 "Issue in running spiderwebs with very different reference times than SFINCS' tref (10s of years)" as in Hurrywave using a higher precision which is already sufficient, not as in branch solution which was more complex

+++ source/src/sfincs_date.f90
@@ -182,9 +182,9 @@ CONTAINS
    ! 
    subroutine time_difference(datespw,datesim,dtsec)
        !
-       integer ijul1, ijul2     
-       integer yyyy1,mm1,dd1,hh1,mn1,ss1,yyyy2,mm2,dd2,hh2,mn2,ss2
-       integer dtsec,sec1,sec2
+       integer*8 ijul1, ijul2     
+       integer*4 yyyy1,mm1,dd1,hh1,mn1,ss1,yyyy2,mm2,dd2,hh2,mn2,ss2
+       integer*8 dtsec,sec1,sec2
        !
        character*15  :: datespw
        character*15  :: datesim
