+++ source/src/sfincs_spiderweb.f90
@@ -7,7 +7,7 @@ contains
    implicit none
    !
    integer j, j2, it, ip, n, m, nquant, stat, nheader, id
-   integer dtsec
+   integer*8 dtsec
    integer iyspw,imspw,idspw
    character*4 cyspw 
    character*2 cmspw 
@@ -145,7 +145,7 @@ contains
    implicit none
    !
    integer j, j2, it, ip, n, m, nquant, stat, nheader, id
-   integer dtsec
+   integer*8 dtsec
    integer iyspw,imspw,idspw
    character*4 cyspw 
    character*2 cmspw 
@@ -473,7 +473,7 @@ contains
    !   
    character(*), intent(in) :: line
    character*15, intent(in) :: trefstr
-   integer, intent(out)     :: dtsec
+   integer*8, intent(out)     :: dtsec
    !
    integer j, j2, iopt
    character*4 cyspw
